Added the class template Assign to the skeletons, allowing move assignment
    SType now declares default copy- and move-assignment operators and a
template operator= declaring a Tp_ &&tmp. The move assignment operator is
called for temp. SType objects, in all other cases the template op= is
called. This template determines whether the received Tp_ && is in fact an
rvalue reference and passes this, together with the raw (unreferenced) Tp_ to
Assign::assign. Assign is a template struct calling
    SType's copy-assignment operator for SType (const) & arguments;
    reset(semantic__(tp))            for other Tp_ (const) & arguments,
    reset(semantic__(move(tp)))      for Tp_ && arguments,
using various specializations.

An earlier attempt to declare the default copy and move assignment as well as
function templates for Tp_ const & and Tp_ && failed, as the Tp_ && takes
precedence over the copy assignment operator.