New skeletons use a completely redesigned parse() function. Many previously defined members have received new names, and the trailing underline convention is more stricly applied.
Also message indexing has been implemented: once a message index
has been defined, and is redefined in nested grammatical constructs
then the original message index is restored once the nested gramm.
construct has been completed, or when parsing returns to the
surrounding construct when a syntactic error is encountered in the
nested construct.